chore: Improve developer experience with documentation and linting (#256)
I initially updated the README to include:

- A better usage example (as it was not straightforward for me to use the lib)
- The `rubocop` command to lint the code

While doing so I realized that `bundle exec rubocop` didn't work due to missing dependencies in the `gemspec`.

Once that was fixed, it turned out that plenty of files were not matching the Rubocop rules. To avoid including hundreds of changes, I've generate a `.rubocop_todo.yml` file that can be fixed later on. And then I added a CI workflow to run `rubocop`.