misc: Reorder resources for readability (#268)
## Context
  
I found out that there were some resources missing from the client. While adding those missing resources I reordered the resources for readability. This PR extracts those changes to have smaller PRs.

## Description

This changes the resource definition in the client to able ordered alphabetically.

When doing so, I found out that the client depends on the resources but actually did not require them so I added the missing `require`.