chore(linting): Improve Rubocop consistency (#287)
## Context

The library currently does not depends on Rails so it doesn't make sense to rely on `rubocop-rails` as it could lead to invalid correction by Rubocop. For instance, it would try to fix `Time.now` by `Time.zone.now` which is not available.

Also, we expect this client to work with Ruby 3.1 but the Rubocop configuration actually had no `TargetRubyVersion` set which could lead to code working locally but breaking on the CI.

## Description

This removes `rubocop-rails` and set `TargetRubyVersion` to 3.1.