test: Add integration tests calling a real API (#286)
## Context

We've noticed that the current unit tests are insufficient to properly ensure the code quality we expect from our client libraries.

## Description

This introduces the concept of integration tests which runs against a real Lago instance.

To simplify the effort necessary to run those tests, I added Docker Compose file which includes a running Lago instance. It is also possible to run those tests against an instance running outside the Docker compose (local or staging environment for instance).